requirements.txt
PySide6
PyMuPDF
filetype
Pillow
python-docx